Initial commit
This project prototype is a variation of the inglorious-brunch project.

The difference here is that we strip off Jade and use React for web
components implementation.

We are also using Reflux to give a highly decoupled yet simple structure,
as stated in http://spoike.ghost.io/deconstructing-reactjss-flux/

Also we're not giving up with CoffeeScript, so even React components are
.cjsx files instead of .jsx! ;)