Use the Reflux.connect sugar
Ok this is the last one. Hopefully. Instead of setting a component state, we
can simply connect to changes in the store and Reflux will implicitely set
the new status. Now this syntax is comparable to Angular, even more because
Rect's components are a much easier and more elegant way to describe a web
component rather than Angular's directives.