Add a Reflux/React test
This is quite new, I haven't found an example like this in all the interwebz.

Using Reflux we can trigger events through actions and inspect stores, which
abstracts a lot from all the 'click on that button and check the DOM' crap.

However if we want to check React components I've added two ways of doing it.